refactor: remove redundant abstractions of SkillsModal
