--- src/components/SkillModal/ISkillModal.ts
@@ -1,6 +1,0 @@
-import {ComponentProps} from 'react'
-import {Skill} from '../../models'
-
-export interface ISkillModal extends ComponentProps<'article'> {
-    skill: Skill
-}
