--- src/components/SkillModalContent/ISkillModalContent.ts
@@ -1,8 +1,0 @@
-import {PropsWithChildren} from 'react'
-
-export interface ISkillModalContent extends PropsWithChildren {
-    title: string
-    description: string
-    bannerUrl: string
-    bannerAlt: string
-}
