--- src/components/SkillModalContent/index.ts
@@ -1,1 +1,0 @@
-export * from './SkillModalContent'
