--- src/components/SkillModalHeader/ISkillModalHeader.ts
@@ -1,6 +1,0 @@
-import {PropsWithChildren} from 'react'
-
-export interface ISkillModalHeader extends PropsWithChildren {
-    title: string
-    category: string
-}
