--- src/components/SkillModalHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './SkillModalHeader'
