+++ src/components/index.ts
@@ -6,7 +6,7 @@ export * from './ScrollTop'
 export * from './Modal'
 export * from './Navbar'
 export * from './ProjectModal'
-export * from './SkillModal'
+export * from './SkillsModal'
 export * from './SkillCard'
 export * from './ProjectCard'
 export * from './ProjectFilter'
